% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{inference}
\alias{inference}
\title{algorithm}
\usage{
inference(
  data,
  gene_name = "gene1",
  algorithm_name = "aracne",
  criterion = "bic",
  threshold = 0.2
)
}
\value{
strength table
}
\description{
Reconstruct the EGRN using Bayesian Network
}
\examples{
hello("Alice")
}
